+++ CHANGELOG.md
@@ -88,7 +88,7 @@
 * Support for pasting text into fuzzel. Both the regular clipboard,
   and the primary clipboard are supported. See the new
   `clipboard-paste` and `primary-paste` key bindings ([#200][200]).
-* `hide-when-prompt-empty` command line and config option, hiding the
+* `hide-before-typing` command line and config option, hiding the
   application list until the first letter is typed. ([#351][#351]).
 
 [344]: https://codeberg.org/dnkl/fuzzel/issues/344